Move code out of main
