+++ c_cxx/accuracy_tool/src/basic_utils.h
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 #pragma once
+#include <algorithm>
 #include <array>
 #include <cassert>
 #include <cmath>
